+++ src/main/resources/db/migration/V00002__update_user_table_with_hasaccess_field.sql
@@ -1,0 +1,3 @@
+-- Update tg_user table
+ALTER TABLE tg_user
+ADD has_google_access BOOLEAN DEFAULT FALSE;